requirements.txt
streamlit>=1.35.0
openai>=1.40.0
python-docx>=1.1.2
reportlab>=4.2.2